+++ database/migrations/2022_04_15_231722_create_commissions_table.php
@@ -16,16 +16,14 @@ return new class extends Migration
         Schema::create('commissions', function (Blueprint $table) {
             $table->engine="InnoDB";
             $table->id();
-            $table->float('amount');
+            $table->double('amount');
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('user_id');
             $table->timestamps();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-
         });
     }
-
     /**
      * Reverse the migrations.
      *
